pipelines/pyproject.toml
[tool.poetry]
name = "rag-pipeline"
version = "0.1.0"
description = "Pipeline for processing markdown documents and uploading them to Milvus for RAG applications"
authors = ["ssgrummons <ssgrummo@us.ibm.com>"]
readme = "README.md"
packages = [{include = "pipeline.py"}]

[tool.poetry.dependencies]
python = "^3.9"
pymilvus = "^2.3.6"
sentence-transformers = "^2.5.1"
keybert = "^0.7.0"
python-frontmatter = "^1.0.0"
markdown = "^3.5.2"
langchain = "^0.1.9"
pyyaml = "^6.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-cov = "^4.1.0"
pytest-mock = "^3.12.0"
black = "^24.1.1"
isort = "^5.13.2"
mypy = "^1.8.0"
pylint = "^3.0.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py39']

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=pipeline --cov-report=term-missing" 